ANP-85 Move API logic from ScanViewModel to repository and implement Koin dependency injection for view model
