+++ android/app/src/main/java/hr/foi/air/wattsup/di/ViewModelsModule.kt
@@ -3,6 +3,7 @@ package hr.foi.air.wattsup.di
 import hr.foi.air.wattsup.viewmodels.AuthenticationViewModel
 import hr.foi.air.wattsup.viewmodels.CardViewModel
 import hr.foi.air.wattsup.viewmodels.HistoryViewModel
+import hr.foi.air.wattsup.viewmodels.ScanViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
@@ -19,6 +20,11 @@ val viewModelsModule = module {
     }
     viewModel {
         HistoryViewModel(
+            repository = get(),
+        )
+    }
+    viewModel {
+        ScanViewModel(
             repository = get(),
         )
     }
